Zmena datoveho typu u Id na int

+++ DopravniPodnik/DopravniPodnik/Data/Models/Adresy.cs
@@ -5,13 +5,13 @@ namespace DopravniPodnik.Data.Models;
 
 public partial class Adresy
 {
-    public decimal IdAdresy { get; set; }
+    public int IdAdresy { get; set; }
 
     public string Mesto { get; set; } = null!;
 
     public string Ulice { get; set; } = null!;
 
-    public short CisloPopisne { get; set; }
+    public int CisloPopisne { get; set; }
 
     public virtual ICollection<Uzivatele> Uzivateles { get; set; } = new List<Uzivatele>();
 }
