+++ DopravniPodnik/DopravniPodnik/Data/Models/Zastaveni.cs
@@ -5,13 +5,13 @@ namespace DopravniPodnik.Data.Models;
 
 public partial class Zastaveni
 {
-    public decimal IdZastaveni { get; set; }
+    public int IdZastaveni { get; set; }
 
     public DateTime Odjezd { get; set; }
 
-    public decimal IdLinky { get; set; }
+    public int IdLinky { get; set; }
 
-    public decimal IdZastavky { get; set; }
+    public int IdZastavky { get; set; }
 
     public virtual Linky IdLinkyNavigation { get; set; } = null!;
